requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
tortoise-orm[asyncpg]>=0.20,<1.0
pyjwt==2.12.1
bcrypt>=5.0.0
python-multipart==0.0.22
python-dotenv==1.2.2
pydantic-settings>=2.0.0
streamlit==1.55.0
minganci>=1.0.0
pytest>=8.0.0,<9.0.0
